sdk/tests/conformance/drawingbuffer-test.html

<!--
Copyright (c) 2009 The Chromium Authors. All rights reserved.
Use of this source code is governed by a BSD-style license that can be
found in the LICENSE file.
 -->
<!doctype html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<title>WebGL Canvas Conformance Tests</title>
<link rel="stylesheet" href="../resources/js-test-style.css"/>
<script src="../resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" style="width: 50px; height: 50px;"> </canvas>
<canvas id="canvas2d" width="40" height="40"> </canvas>
<script>
if (window.initNonKhronosFramework) {
  window.initNonKhronosFramework(true);
}

description("This test ensures WebGL implementations interact correctly implement drawingbufferWidth/Height.");

debug("");

var err;
var maxSize;
var canvas = document.getElementById("canvas");
var canvas2d = document.getElementById("canvas2d");
var ctx2d = canvas2d.getContext("2d");
var gl = create3DContext(canvas);
if (!gl) {
  testFailed("context does not exist");
} else {
  testPassed("context exists");

  debug("");
  debug("Checking drawingBufferWidth/drawingBufferHeight");

  // Check that a canvas with no width or height is 300x150 pixels
  shouldBe('gl.drawingBufferWidth', 'canvas.width');
  shouldBe('gl.drawingBufferHeight', 'canvas.height');

  // Check changing the canvas size to something thattoo large falls back to resonable values.
  maxSize = gl.getParameter(gl.MAX_VIEWPORT_DIMS);
  shouldBeTrue('maxSize[0] > 0');
  shouldBeTrue('maxSize[1] > 0');

  debug("MAX_VIEWPORT_DIMS = " + maxSize[0] + "x" + maxSize[1]);
  canvas.width = maxSize[0] * 4;
  canvas.height = maxSize[1] * 4;
  shouldBeTrue('canvas.drawingBufferWidth > 0');
  shouldBeTrue('canvas.drawingBufferHeight > 0');
}
debug("")
successfullyParsed = true;
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
